+++ chancy/plugins/workflow/django/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ChancyWorkflowConfig(AppConfig):
+    name = "chancy.plugins.workflow.django"
+    label = "chancy_workflow"
+    verbose_name = "Chancy Workflow"