+++ tests/contrib/django/settings.py
@@ -1,0 +1,7 @@
+"""
+Django test settings for chancy.contrib.django tests.
+"""
+
+INSTALLED_APPS = [
+    "chancy.contrib.django.apps.ChancyConfig",
+]